task answer

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -7,7 +7,7 @@
  */
 int get_bit(unsigned long int n, unsigned int index)
 {
-	unsigned long int i = 1 << index;
+	int i = 1 << index;
 	int j = (n & i) >> index;
 
 	if (i > n || index >= sizeof(unsigned long int) * 8)
